Add (untested, no h/w) 10160 support. Propagate some spelling fixes. Add back in merged 'mhz' change.
Use A64 variants to initialize ULTRA2 or better && 1240 cards. If PAE
is ever implemented for *BSD, we'll be closer to being ready to go.

Add in a commented out example for how to adjust idelaytimer on 2300s.

If we have ISP_FW_CRASH_DUMP defined, then don't reinit the isp in
ISP_CONN_FATAL case- that's up to the platform outer layer to do after
it finishes saving a f/w crashdump.

Replaces STRNCAT with SNPRINTF.