Delete the password from the ftp url shown when no sets are found.
a) We shouldn't display passwords.

b) If it is encoded, and thus different from what was entered,
installers will leap to the incorrect conclusion that the logon
failed.

c) At this point the logon to the server was successful so the
password worked, and the user shouldn't have forgotten it already.

ok deraadt@.