very unlikely overflow.  but sticking to the idiom is important:  thereby, example by example, we teach people how to actually use snprintf.  because it is clear (especially judging by code coming from netbsd hint hint perhaps if i say it like this they will finally learn) that people are not paying attention, and replacing one security problem with another.
in the early days we replaced buffer the typical ANSI-C standardized function
buffer overflows (by which I mean strcpy, strcat, and sprintf) with
non-overflowing ones -- range checking varients.  We knew we were fixing
a major problem.  The damn overflows.  But we did not have time in all cases
to handle the next problem we were not handling:  string truncation.  Now we
need to (I hope not slowly) start fixing the string truncations.

Anyone going to help?