rename tht_rx to tht_rxf_fill to avoid confusion.
keep a list of used pkts in tht_pkt_list as well as free ones. add
tht_pkt_used which can be used to see if a tht_pkt_list is in use by
returning the head of the used list inside itself.

(ab)use this to drain the rx free fifo when the interface is brought down.
tht_rxf_drain pulls used pkts off, unmaps their mbufs, frees them, and then
returns the pkt to the free list. the whole tht_pkt_list can then be freed
safely.