more places where the difference between the hardware sample size and the userland sample size matters.
there may be different sample size factors for play and record,
so use the appropriate factor when userland sets the blocksize.

the reported blocksize, hiwat and lowat are only for the play buffer.
IMO, it would be good to report the record blocksize, hiwat and lowat
as well, but that is another issue.

tested with emu(4) and an auvia(4) modified to only do stereo in
hardware.