Blind support for SGI PIC PCI-X controller found on Origin 350 and Tezro systems. PIC was supposed to be mostly XBridge compatible, but a silicon bug prevents it from working correctly if 32 bit register writes are used, so the xbridge(4) code now needs to issue 64 bit writes.
In order to make this a bit more transparent, rebase all widget registers to
a 8 byte boundary, and provide a few inline accessors which will do the right
thing if you want 32 bit writes.

Tested to not cause regressions on Octane, Origin 200 and Fuel; covering
Bridge < 4, Bridge >= 4 and XBridge flavours.