the ata identify structure is stupid. its full of buffers you have to swap to use.
this diff doesnt swap fields the identify structure when its fetched, but
adds a function that swaps them and copies them into a destination buffer.
there is no functional change to the existing users of these fields,
specifically the scsi inquiry faker and the device serial vpd page.

this does fix the devid vpd page though. if a disk has a wwn we now present
it correctly.