cut mfi over to using iopools.
this gets rid of NO_CCB, provides better scheduling of io between busy
logical volumes on the same controller, and makes the ioctl and sensor
paths more reliable.

this is the least intrusive version of this change i could come up with. it
only took 4 or 5 goes...

ok marco@
deraadt@ says this is a good time for this change.