Do not complain about an /etc/group line "+\n" as "wrong number of fields", that abbreviated syntax is explicitly allowed by group(5).  While here, warn if it isn't the last line in the group file. Regression reported, fix tested and ok miod@, and seems good to ajacoutot@.
Note: I'm not removing the advice to put "+\n" at the end of the group file
right now because i'm not 100% sure that advice is pointless, even though
guenther@ looked at the code an came to the conclusion OpenBSD libc ought
to cope.  And i'd rather have the manuals and the syntax checker be
consistent.  In case this really annoys people, it can be carefully tested
and changed later.