libexec/tftpd handled the case where we'd get an ack for the previous block by flushing the data on the socket and waiting for a retransmit by timeout to occur.
my stuff (usr.sbin/tftpd) had XXXs and failure in those places.

this diff fixes that.

this should address the problems that sthen and jcs have been having
(and which i was finally able to reproduce here).

it also avoids reusing the clients buffer to both send and recv
frames. we recv onto the stack now so retry always sends what we
originally built for the peer.

tested by and ok jcs@ sthen@