Loongson: Add a stat clock for the lemote machines.
The stat clock is implemented so that it also covers, at least
according to the tests I've done, Torek's paper on randomized sampling.

Most of the bits for randomizing the stat ticks were taken from sparc's
implementation and adapted to the companion chip's MFGPT frequencies.

Before this my lemote showed a 8% cpu usage for the cpuhog example from
Torek's paper even though openssl speed showed differences up to 27%
when ran with and without cpuhog. With the new stat clock  it shows
cpuhog around 18-22%, which I think is the proper value.

Okay miod@.